Scanner starting to work - parsefiles seems ok, monitor not

+++ EDDiscovery/EliteDangerous/NetLogClass.cs
@@ -59,9 +59,9 @@ namespace EDDiscovery
 
             List<JournalLocOrJump> visitedSystems = new List<JournalLocOrJump>();
             Dictionary<string, TravelLogUnit> m_travelogUnits = TravelLogUnit.GetAll().Where(t => t.type == 1).GroupBy(t => t.Name).Select(g => g.First()).ToDictionary(t => t.Name);
-            Dictionary<int, string> travellogunitid2name = m_travelogUnits.Values.ToDictionary(t => (int)t.id, t => t.Name);
+            Dictionary<long, string> travellogunitid2name = m_travelogUnits.Values.ToDictionary(t => t.id, t => t.Name);
             Dictionary<string, List<JournalLocOrJump>> vsc_lookup = JournalEntry.GetAll().OfType<JournalLocOrJump>().GroupBy(v => v.JournalId).Where(g => travellogunitid2name.ContainsKey(g.Key)).ToDictionary(g => travellogunitid2name[g.Key], g => g.ToList());
-            HashSet<int> journalids = new HashSet<int>(m_travelogUnits.Values.Select(t => (int)t.id).ToList());
+            HashSet<long> journalids = new HashSet<long>(m_travelogUnits.Values.Select(t => t.id).ToList());
             List<JournalLocOrJump> vsSystemsList = JournalEntry.GetAll(currentcmdrid).OfType<JournalLocOrJump>().Where(j => journalids.Contains(j.JournalId)).ToList();
 
             if (vsSystemsList != null)
